DSND-2420: Move status properties to model package
