+++ src/test/java/uk/gov/companieshouse/filinghistory/api/service/FilingHistoryStatusServiceTest.java
@@ -11,8 +11,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.companieshouse.filinghistory.api.service.CompanyNumberStatusProcessor.CompanyNumberAffixes;
-import uk.gov.companieshouse.filinghistory.api.statusrules.PrefixProperties;
-import uk.gov.companieshouse.filinghistory.api.statusrules.StatusRuleProperties;
+import uk.gov.companieshouse.filinghistory.api.model.statusrules.PrefixProperties;
+import uk.gov.companieshouse.filinghistory.api.model.statusrules.StatusRuleProperties;
 
 @ExtendWith(MockitoExtension.class)
 class FilingHistoryStatusServiceTest {
